requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
pytrends==4.9.2
jinja2==3.1.2
pydantic==2.5.2
certifi==2023.11.17 